Refactor search functionality: move highlight and mapHref utilities to search.tsx and clean up SearchOverlay component
